Extract Parser class and CommandType enum
