+++ src/main/java/Parser.java
@@ -1,2 +1,60 @@
-package PACKAGE_NAME;public class Parser {
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
+
+public class Parser {
+
+    public static CommandType parseCommand(String input) {
+        if (input.equalsIgnoreCase("bye")) return CommandType.BYE;
+        if (input.equalsIgnoreCase("list")) return CommandType.LIST;
+        if (input.startsWith("todo")) return CommandType.TODO;
+        if (input.startsWith("deadline")) return CommandType.DEADLINE;
+        if (input.startsWith("event")) return CommandType.EVENT;
+        if (input.startsWith("mark")) return CommandType.MARK;
+        if (input.startsWith("unmark")) return CommandType.UNMARK;
+        if (input.startsWith("delete")) return CommandType.DELETE;
+        if (input.startsWith("tasks on")) return CommandType.TASKS_ON_DATE;
+        return CommandType.UNKNOWN;
+    }
+
+    private static Task parseFileTask(String taskData) throws DukeException {
+        String[] parts = taskData.split(" \\| ");
+        switch (parts[0]) {
+            case "T":
+                ToDo todo = new ToDo(parts[2]);
+                if (parts[1].equals("1")) todo.markAsDone();
+                return todo;
+            case "D":
+                Deadline deadline = new Deadline(parts[2], parts[3]);
+                if (parts[1].equals("1")) deadline.markAsDone();
+                return deadline;
+            case "E":
+                Event event = new Event(parts[2], parts[3], parts[4]);
+                if (parts[1].equals("1")) event.markAsDone();
+                return event;
+            default:
+                throw new DukeException("Unknown task type: " + parts[0]);
+        }
+    }
+
+    public static LocalDate getLocalDate(String input) throws DukeException {
+        String[] dateParts = input.split(" ");
+        if (dateParts.length < 3) {
+            throw new DukeException("Please provide a valid date in the format d/M/yyyy.");
+        }
+        LocalDate givenDate;
+        try {
+            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("d/M/yyyy");
+            givenDate = LocalDate.parse(dateParts[2], formatter);
+        } catch (DateTimeParseException e) {
+            throw new DukeException("Invalid date format. Please use d/M/yyyy.");
+        }
+        return givenDate;
+    }
+
+    public static boolean isWithinEventDate(Event event, LocalDate date) {
+        LocalDate startDate = event.getFrom().toLocalDate();
+        LocalDate endDate = event.getTo().toLocalDate();
+        return (date.isEqual(startDate) || date.isEqual(endDate) || (date.isAfter(startDate) && date.isBefore(endDate)));
+    }
 }
